changed dish amount to float

+++ src/main/kotlin/foodPlaner/Ingredient.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Ingredient(
-    val amount: Int,
+    val amount: Float,
     val unit: String,
     val name: String
 )
